requirements.txt
# requirements.txt — v2
# Instalar con: pip install -r requirements.txt

# ── Telegram ──
python-telegram-bot[job-queue]==21.5

# ── APIs de datos ──
requests==2.31.0
understatapi==0.9.0

# ── Análisis de datos ──
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# ── Scraping (Rushbet) ──
playwright==1.41.2
# Después de instalar, ejecutar: playwright install chromium

# ── Google Sheets (base de datos) ──
gspread==6.0.2
google-auth==2.27.0

# ── Imágenes ──
Pillow==10.2.0

# ── Variables de entorno ──
python-dotenv==1.0.1

# ── Zona horaria ──
pytz==2024.1